requirements.txt
numpy==1.24.3
opencv-python
Pillow
tensorflow-gpu==2.10.0
matplotlib
scikit-learn
seaborn